Re-org error modal
